+++ completion/git-machete.completion.bash
@@ -10,15 +10,15 @@ _git_machete() {
     common_opts="--debug -h --help -v --verbose --version"
     add_opts="-o --onto="
     diff_opts="-s --stat"
-    discover_opts="-l --list-commits -r --roots="
+    discover_opts="-C --checked-out-since= -l --list-commits -r --roots="
     reapply_opts="-f --fork-point="
     slide_out_opts="-d --down-fork-point="
     status_opts="-l --list-commits --color="
     traverse_opts="-l --list-commits"
-    traverse_opts="-l --list-commits"
     update_opts="-f --fork-point="
 
     case "$cur" in
+        --checked-out-since=*) __gitcomp "" ;;
         --down-fork-point=*|--fork-point=*) __gitcomp "$(__git_refs)" "" "${cur##--*=}" ;;
         --onto=*) __gitcomp_nl "$(git machete list managed)" "" "${cur##--onto=}" ;;
         --roots=*) __gitcomp "$(__git_heads)" "" "${cur##--roots=}" ;;
@@ -41,6 +41,7 @@ _git_machete() {
              else
                 prev="${COMP_WORDS[COMP_CWORD-1]}"
                 case "$prev" in
+                    -C|--checked-out-since) __gitcomp "" ;;
                     -d|--down-fork-point|-f|--fork-point) __gitcomp "$(__git_refs)" ;;
                     # TODO #25: We don't complete --help since it's going to be captured by git anyway
                     # (and results in redirection to yet non-existent man for `git-machete`).
@@ -66,7 +67,7 @@ _git_machete() {
                                 else
                                     __gitcomp_nl "$(git machete list slidable-after "$prev" 2>/dev/null)"
                                 fi ;;
-                            *) COMPREPLY=() ;;
+                            *) __gitcomp "" ;;
                         esac ;;
                 esac
             fi